Combine path- and versionfinder packages
