+++ src/debug/debug.go
@@ -3,11 +3,11 @@ package debug
 import (
 	"os"
 
-	"github.com/anttiharju/vmatch-golangci-lint/src/pathfinder"
+	"github.com/anttiharju/vmatch-golangci-lint/src/finder"
 )
 
 func getFilePath() string {
-	return pathfinder.GetBinDir() + string(os.PathSeparator) + "debug.txt"
+	return finder.GetBinDir() + string(os.PathSeparator) + "debug.txt"
 }
 
 func WriteToFile(s string) {
