+++ src/finder/finder.go
@@ -1,4 +1,4 @@
-package versionfinder
+package finder
 
 import (
 	"os"
@@ -8,11 +8,35 @@ import (
 
 	"github.com/anttiharju/vmatch-golangci-lint/src/exit"
 	"github.com/anttiharju/vmatch-golangci-lint/src/exit/exitcode"
-	"github.com/anttiharju/vmatch-golangci-lint/src/pathfinder"
 )
 
+func GetBinDir() string {
+	binPath := getBin()
+	binDir := binPath[:strings.LastIndex(binPath, string(os.PathSeparator))]
+
+	return binDir
+}
+
+func getBin() string {
+	binPath, err := os.Executable()
+	if err != nil {
+		exit.WithMessage(exitcode.BinPathIssue, "Cannot get executable path")
+	}
+
+	return binPath
+}
+
+func GetWorkDir() string {
+	workdir, err := os.Getwd()
+	if err != nil {
+		exit.WithMessage(exitcode.WorkDirIssue, "Cannot get working directory")
+	}
+
+	return workdir
+}
+
 func GetVersion(filename string) string {
-	workDir := pathfinder.GetWorkDir()
+	workDir := GetWorkDir()
 
 	for {
 		filePath := filepath.Join(workDir, filename)
